- remove unused RegistrationController and RegistrationControllerTest - allow user registration in security config - add login endpoint - add findByEmail in UserService - moved tests from RegistrationControllerTest to UserControllerTest
