+++ src/main/java/software/sigma/internship/controller/UserController.java
@@ -7,13 +7,17 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.i18n.LocaleContextHolder;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import software.sigma.internship.dto.AuthUserDto;
 import software.sigma.internship.dto.UserDto;
 import software.sigma.internship.dto.UserFullDto;
 import software.sigma.internship.service.UserService;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @Tag(name = "User controller")
@@ -35,12 +39,21 @@ public class UserController {
     }
 
     @Operation(summary = "Register new user")
+    @ResponseStatus(HttpStatus.CREATED)
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "User is created successfully")
     })
     @PostMapping
-    public UserDto registration(@RequestBody AuthUserDto user) {
+    public UserDto registration(@Valid @RequestBody AuthUserDto user) {
         log.info(LocaleContextHolder.getLocale().getLanguage());
         return userService.register(user, LocaleContextHolder.getLocale().getLanguage());
+    }
+    @PreAuthorize("hasAnyAuthority('USER', 'ADMIN')")
+    @GetMapping("/login")
+    public UserFullDto authenticatedUser() {
+        UserDetails authenticatedUser = (UserDetails) SecurityContextHolder.getContext()
+                .getAuthentication()
+                .getPrincipal();
+        return userService.findByEmail(authenticatedUser.getUsername());
     }
 }
