+++ src/main/java/software/sigma/internship/service/UserService.java
@@ -7,6 +7,7 @@ import software.sigma.internship.dto.UserFullDto;
 import java.util.List;
 
 public interface UserService {
+    UserFullDto findByEmail(String email);
     UserDto register(AuthUserDto user, String isoCode);
 
     boolean checkIfUserExist(String email);
